+++ CHANGELOG.rst
@@ -82,6 +82,7 @@ Fixed
 - Fixed :issue:`543`, :issue:`552`, :issue:`553` - bugs with various Query handlings
 - Fixed :issue:`575` - cannot access maps due to bug in login and authentication in Interaction class
 - Fixed :issue:`539` - print downloadList target directory
+- Fixed :issue:`566` - made error message for web query with non-unique parameters name more specific
 
 Refactored
 ^^^^^^^^^^